add versioned models and schemas
