+++ internal/provider/task_resource.go
@@ -6,23 +6,16 @@ package provider
 import (
 	"context"
 	"fmt"
-
+	v0_models "terraform-provider-tasks/internal/provider/models/v0"
+	v1_models "terraform-provider-tasks/internal/provider/models/v1"
+	v1_schemas "terraform-provider-tasks/internal/provider/schemas/v1"
 	"terraform-provider-tasks/internal/sdk"
-	entitites "terraform-provider-tasks/internal/sdk/entities"
+	entities "terraform-provider-tasks/internal/sdk/entities"
 
-	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema/booldefault"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringdefault"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
-	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 )
-
-const TASK_SCHEMA_VERSION = 1
 
 // Ensure provider defined types fully satisfy framework interfaces.
 var _ resource.Resource = &TaskResource{}
@@ -37,208 +30,12 @@ type TaskResource struct {
 	client *sdk.TasksClient
 }
 
-type InputProcessorsModel struct {
-	InputProcessors []InputProcessorModel `tfsdk:"input_processor"`
-}
-
-type InputProcessorModel struct {
-	ParamName      types.String            `tfsdk:"param_name"`
-	InputProcessor types.String            `tfsdk:"input_processor"`
-	Config         map[string]types.String `tfsdk:"config"`
-}
-
-// TaskResourceModel describes the resource data model.
-type TaskResourceModel struct {
-	ActiveRevisionID types.String                             `tfsdk:"active_revision_id"`
-	Description      types.String                             `tfsdk:"description"`
-	Enabled          types.Bool                               `tfsdk:"enabled"`
-	ExposedToAgents  types.Bool                               `tfsdk:"exposed_to_agents"`
-	ID               types.String                             `tfsdk:"id"`
-	ImageRequired    types.Bool                               `tfsdk:"image_required"`
-	InputProcessors  *InputProcessorsModel                    `tfsdk:"input_processors"`
-	LLMModelID       types.String                             `tfsdk:"llm_model_id"`
-	Name             types.String                             `tfsdk:"name"`
-	OptimiseImages   types.Bool                               `tfsdk:"optimise_images"`
-	OutputFormat     map[string]entitites.OutputFormatWrapper `tfsdk:"output_format"`
-	OutputModality   types.String                             `tfsdk:"output_modality"`
-	Public           types.Bool                               `tfsdk:"public"`
-	SystemPrompt     types.String                             `tfsdk:"system_prompt"`
-	UserPrompt       types.String                             `tfsdk:"user_prompt"`
-}
-
-func (trm *TaskResourceModel) HasInputProcessors() bool {
-	return trm.InputProcessors != nil && len(trm.InputProcessors.InputProcessors) > 0
-}
-
-func (trm *TaskResourceModel) PopulateFromTaskEntity(task *entitites.Task) error {
-
-	rev, err := task.GetActiveRevision()
-	if err != nil {
-		return err
-	}
-
-	trm.ID = types.StringValue(task.ID)
-	trm.Name = types.StringValue(task.Name)
-	trm.Enabled = types.BoolValue(task.Enabled)
-	trm.ExposedToAgents = types.BoolValue(task.ExposedToAgents)
-	trm.Public = types.BoolValue(task.Public)
-	trm.Description = types.StringValue(task.Description)
-
-	trm.OptimiseImages = types.BoolValue(rev.OptimiseImages)
-	trm.SystemPrompt = types.StringValue(rev.SystemPrompt)
-	trm.UserPrompt = types.StringValue(rev.UserPrompt)
-	trm.LLMModelID = types.StringValue(rev.LLMModelID)
-	trm.ImageRequired = types.BoolValue(rev.ImageRequired)
-	trm.OptimiseImages = types.BoolValue(rev.OptimiseImages)
-
-	if rev.HasInputProcessors() {
-		trm.InputProcessors = &InputProcessorsModel{
-			InputProcessors: make([]InputProcessorModel, len(*rev.InputProcessors)),
-		}
-		for i, ip := range *rev.InputProcessors {
-			trm.InputProcessors.InputProcessors[i] = InputProcessorModel{
-				ParamName:      types.StringValue(ip.ParamName),
-				InputProcessor: types.StringValue(ip.InputProcessor),
-				Config:         make(map[string]types.String, len(ip.Config)),
-			}
-			for k, v := range ip.Config {
-				trm.InputProcessors.InputProcessors[i].Config[k] = types.StringValue(v)
-			}
-		}
-	}
-
-	trm.ActiveRevisionID = types.StringValue(rev.ID)
-
-	return nil
-}
-
 func (r *TaskResource) Metadata(ctx context.Context, req resource.MetadataRequest, resp *resource.MetadataResponse) {
 	resp.TypeName = req.ProviderTypeName + "_task"
 }
 
 func (r *TaskResource) Schema(ctx context.Context, req resource.SchemaRequest, resp *resource.SchemaResponse) {
-	resp.Schema = schema.Schema{
-
-		MarkdownDescription: "Task resource",
-		Version:             TASK_SCHEMA_VERSION,
-
-		Attributes: map[string]schema.Attribute{
-			"id": schema.StringAttribute{
-				Computed:            true,
-				MarkdownDescription: "Identifier",
-				PlanModifiers: []planmodifier.String{
-					stringplanmodifier.UseStateForUnknown(),
-				},
-			},
-			"name": schema.StringAttribute{
-				Required:    true,
-				Description: "A name or reference for the Task.",
-			},
-			"description": schema.StringAttribute{
-				Optional:    true,
-				Description: "A description of the Task.",
-			},
-			"enabled": schema.BoolAttribute{
-				Required:    true,
-				Description: "When `true` the Task is active and callable.",
-			},
-			"public": schema.BoolAttribute{
-				Optional:    true,
-				Description: "",
-				Default:     booldefault.StaticBool(false),
-				Computed:    true,
-			},
-			"exposed_to_agents": schema.BoolAttribute{
-				Optional:    true,
-				Description: "",
-				Default:     booldefault.StaticBool(false),
-				Computed:    true,
-			},
-			"active_revision_id": schema.StringAttribute{
-				Computed: true,
-			},
-			// revision specific
-			"system_prompt": schema.StringAttribute{
-				Required:    true,
-				Description: "The system prompt that is used to set the LLM context.",
-			},
-			"user_prompt": schema.StringAttribute{
-				Required:    true,
-				Description: "The user prompt that is used to set the LLM context.",
-			},
-			"llm_model_id": schema.StringAttribute{
-				Required:    true,
-				Description: "The ID of the LLM model to use for the Task.",
-			},
-			"image_required": schema.BoolAttribute{
-				Optional:    true,
-				Description: "When true it requires an image to be sent in the Task Run request.",
-				Default:     booldefault.StaticBool(false),
-				Computed:    true,
-			},
-			"optimise_images": schema.BoolAttribute{
-				Optional:    true,
-				Description: "When true (default) images will be automatically optimised before processing. Set to false to disable lossy image optimisation.",
-				Default:     booldefault.StaticBool(true),
-				Computed:    true,
-			},
-			"output_modality": schema.StringAttribute{
-				Optional:    true,
-				Description: "Specifies the output modality of the task. Can be 'json' or 'image'",
-				Default:     stringdefault.StaticString("json"),
-				Computed:    true,
-				Validators: []validator.String{
-					stringvalidator.OneOf("json", "image"),
-				},
-			},
-			"output_format": schema.MapNestedAttribute{
-				Required: true,
-				NestedObject: schema.NestedAttributeObject{
-					Attributes: map[string]schema.Attribute{
-						"type": schema.StringAttribute{
-							Required: true,
-						},
-						"description": schema.StringAttribute{
-							Optional: true,
-						},
-						"item_type": schema.StringAttribute{
-							Optional: true,
-						},
-						"object": schema.MapAttribute{
-							Optional:    true,
-							ElementType: types.StringType,
-						},
-						"options": schema.MapAttribute{
-							Optional:    true,
-							ElementType: types.StringType,
-						},
-					},
-				},
-			},
-		},
-		Blocks: map[string]schema.Block{
-			"input_processors": schema.SingleNestedBlock{
-				Blocks: map[string]schema.Block{
-					"input_processor": schema.ListNestedBlock{
-						NestedObject: schema.NestedBlockObject{
-							Attributes: map[string]schema.Attribute{
-								"param_name": schema.StringAttribute{
-									Required: true,
-								},
-								"input_processor": schema.StringAttribute{
-									Required: true,
-								},
-								"config": schema.MapAttribute{
-									Optional:    true,
-									ElementType: types.StringType,
-								},
-							},
-						},
-					},
-				},
-			},
-		},
-	}
+	resp.Schema = v1_schemas.TASK_SCHEMA
 }
 
 func (r *TaskResource) Configure(ctx context.Context, req resource.ConfigureRequest, resp *resource.ConfigureResponse) {
@@ -262,7 +59,7 @@ func (r *TaskResource) Configure(ctx context.Context, req resource.ConfigureRequ
 }
 
 func (r *TaskResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
-	var data TaskResourceModel
+	var data v1_models.TaskResourceModel
 
 	// Read Terraform plan data into the model
 	resp.Diagnostics.Append(req.Plan.Get(ctx, &data)...)
@@ -303,7 +100,7 @@ func (r *TaskResource) Create(ctx context.Context, req resource.CreateRequest, r
 }
 
 func (r *TaskResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
-	var data TaskResourceModel
+	var data v1_models.TaskResourceModel
 
 	// Read Terraform prior state data into the model
 	resp.Diagnostics.Append(req.State.Get(ctx, &data)...)
@@ -327,7 +124,7 @@ func (r *TaskResource) Read(ctx context.Context, req resource.ReadRequest, resp 
 }
 
 func (r *TaskResource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
-	var data TaskResourceModel
+	var data v1_models.TaskResourceModel
 
 	// Read Terraform plan data into the model
 	resp.Diagnostics.Append(req.Plan.Get(ctx, &data)...)
@@ -368,7 +165,7 @@ func (r *TaskResource) Update(ctx context.Context, req resource.UpdateRequest, r
 }
 
 func (r *TaskResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
-	var data TaskResourceModel
+	var data v1_models.TaskResourceModel
 
 	// Read Terraform prior state data into the model
 	resp.Diagnostics.Append(req.State.Get(ctx, &data)...)
@@ -388,13 +185,13 @@ func (r *TaskResource) ImportState(ctx context.Context, req resource.ImportState
 	resource.ImportStatePassthroughID(ctx, path.Root("id"), req, resp)
 }
 
-func (r *TaskResource) FormatInputProcessors(data TaskResourceModel) *[]entitites.InputProcessor {
-	ips := []entitites.InputProcessor{}
+func (r *TaskResource) FormatInputProcessors(data v1_models.TaskResourceModel) *[]entities.InputProcessor {
+	ips := []entities.InputProcessor{}
 	if !data.HasInputProcessors() {
 		return &ips
 	}
 	for _, v := range data.InputProcessors.InputProcessors {
-		ip := entitites.InputProcessor{
+		ip := entities.InputProcessor{
 			ParamName:      v.ParamName.ValueString(),
 			InputProcessor: v.InputProcessor.ValueString(),
 		}
@@ -407,4 +204,40 @@ func (r *TaskResource) FormatInputProcessors(data TaskResourceModel) *[]entitite
 		ips = append(ips, ip)
 	}
 	return &ips
+}
+
+func (r *TaskResource) UpgradeState(ctx context.Context) map[int64]resource.StateUpgrader {
+	return map[int64]resource.StateUpgrader{
+		// State upgrade implementation from 0 (prior state version) to 1 (Schema.Version)
+		0: {
+			PriorSchema: &v1_schemas.TASK_SCHEMA,
+			StateUpgrader: func(ctx context.Context, req resource.UpgradeStateRequest, resp *resource.UpgradeStateResponse) {
+
+				var priorStateData = v0_models.TaskResourceModel{
+					OutputFormat: make(map[string]types.String, 0),
+				}
+
+				resp.Diagnostics.Append(req.State.Get(ctx, &priorStateData)...)
+
+				if resp.Diagnostics.HasError() {
+					return
+				}
+
+				// Upgrade logic
+				newOutputFormat := make(map[string]entities.OutputFormatWrapper)
+				for key, val := range priorStateData.OutputFormat {
+					ofs := entities.OutputFormatSimple(val.ValueString())
+					newOutputFormat[key] = entities.OutputFormatWrapper{
+						Simple: &ofs,
+					}
+				}
+
+				upgradedStateData := v1_models.TaskResourceModel{
+					OutputFormat: newOutputFormat,
+				}
+
+				resp.Diagnostics.Append(resp.State.Set(ctx, upgradedStateData)...)
+			},
+		},
+	}
 }
