+++ internal/sdk/entities/task.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package entitites
+package entities
 
 import (
 	"fmt"
